# contracts/openapi/aieos-v1.json
{
  "components": {
    "schemas": {
...
      "ContentResponse": {
        "additionalProperties": false,
        "properties": {
          "aggregate_revision": {
            "title": "Aggregate Revision",
            "type": "integer"
          },
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "content_id": {
            "format": "uuid",
            "title": "Content Id",
            "type": "string"
          },
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Version Id"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "locale": {
            "title": "Locale",
            "type": "string"
          },
          "published_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published Version Id"
          },
          "stewardship_state": {
            "title": "Stewardship State",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "content_id",
          "content_type",
          "title",
          "description",
          "locale",
          "stewardship_state",
          "current_version_id",
          "published_version_id",
          "aggregate_revision",
          "created_at",
          "updated_at",
          "archived_at"
        ],
        "title": "ContentResponse",
        "type": "object"
      },
...
      "ProblemDetails": {
        "$defs": {
          "ProblemErrorItem": {
            "additionalProperties": false,
            "properties": {
              "code": {
                "title": "Code",
                "type": "string"
              },
              "detail": {
                "title": "Detail",
                "type": "string"
              },
              "pointer": {
                "title": "Pointer",
                "type": "string"
              }
            },
            "required": [
              "pointer",
              "code",
              "detail"
            ],
            "title": "ProblemErrorItem",
            "type": "object"
          }
        },
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "format": "uuid",
            "title": "Correlation Id",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProblemErrorItem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errors"
          },
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "request_id",
          "correlation_id"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
      "ProblemErrorItem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "pointer": {
            "title": "Pointer",
            "type": "string"
          }
        },
        "required": [
          "pointer",
          "code",
          "detail"
        ],
        "title": "ProblemErrorItem",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "AIEOS HTTP foundation (GCI-I06). Content create, version append, and review mutations are development/test foundations only and MUST NOT be authorized for production until transactional outbox and required audit-intent persistence are integrated.",
    "title": "AIEOS HTTP API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/contents/{content_id}": {
      "get": {
        "operationId": "content_get",
        "parameters": [
          {
            "in": "path",
            "name": "content_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Content Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-AIEOS-Tenant-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aieos-Tenant-Id"
            }
          },
          {
            "description": "Optional correlation identifier. Malformed values are rejected. The server generates a UUIDv7 when absent. Distinct from request_id.",
            "in": "header",
            "name": "X-AIEOS-Correlation-ID",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentResponse"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "ETag": {
                "description": "Opaque aggregate revision validator",
                "schema": {
                  "type": "string"
                }
              },
              "X-AIEOS-Correlation-ID": {
                "description": "Correlation identifier preserved or generated by the server.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-AIEOS-Request-ID": {
                "description": "Server-generated UUIDv7 request identifier. Callers cannot choose it.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 Problem Details",
            "headers": {
              "X-AIEOS-Correlation-ID": {
                "description": "Correlation identifier preserved or generated by the server.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-AIEOS-Request-ID": {
                "description": "Server-generated UUIDv7 request identifier. Callers cannot choose it.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 Problem Details",
            "headers": {
              "X-AIEOS-Correlation-ID": {
                "description": "Correlation identifier preserved or generated by the server.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-AIEOS-Request-ID": {
                "description": "Server-generated UUIDv7 request identifier. Callers cannot choose it.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 Problem Details",
            "headers": {
              "X-AIEOS-Correlation-ID": {
                "description": "Correlation identifier preserved or generated by the server.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-AIEOS-Request-ID": {
                "description": "Server-generated UUIDv7 request identifier. Callers cannot choose it.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 Problem Details",
            "headers": {
              "X-AIEOS-Correlation-ID": {
                "description": "Correlation identifier preserved or generated by the server.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-AIEOS-Request-ID": {
                "description": "Server-generated UUIDv7 request identifier. Callers cannot choose it.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 Problem Details",
            "headers": {
              "X-AIEOS-Correlation-ID": {
                "description": "Correlation identifier preserved or generated by the server.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-AIEOS-Request-ID": {
                "description": "Server-generated UUIDv7 request identifier. Callers cannot choose it.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 Problem Details",
            "headers": {
              "X-AIEOS-Correlation-ID": {
                "description": "Correlation identifier preserved or generated by the server.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-AIEOS-Request-ID": {
                "description": "Server-generated UUIDv7 request identifier. Callers cannot choose it.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "RFC 9457 Problem Details",
            "headers": {
              "X-AIEOS-Correlation-ID": {
                "description": "Correlation identifier preserved or generated by the server.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              },
              "X-AIEOS-Request-ID": {
                "description": "Server-generated UUIDv7 request identifier. Callers cannot choose it.",
                "schema": {
                  "format": "uuid",
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Content Get",
        "tags": [
          "contents"
        ]
      }
    },
...
  }
}